requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
twilio==8.4.0
Unidecode==1.3.6
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
